// packages/noodl-git/src/core/models/clone-options.ts
/** Additional arguments to provide when cloning a repository */
export type CloneOptions = {
  /** The branch to checkout after the clone has completed. */
  readonly branch?: string;
  /** The default branch name in case we're cloning an empty repository. */
  readonly defaultBranch?: string;
  /** Clone a single branch; this is just for testing. */
  readonly singleBranch?: boolean;
};